# src/lib/server/openapi.product.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Maildrill Product API",
    "version": "0.1.0",
    "description": "Subscribers, lists, segments, tags, templates, suppressions, auth, and campaign send. Auth: x-api-key or Bearer JWT."
  },
  "components": {
    "securitySchemes": {
      "apiKey": { "type": "apiKey", "in": "header", "name": "x-api-key" },
      "bearer": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "paths": {
...
    "/v1/lists": {
      "post": {
        "summary": "Create a list",
        "tags": ["Lists"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "minLength": 1 },
                  "description": { "type": "string", "nullable": true },
                  "color": { "type": "string", "nullable": true }
                },
                "required": ["name"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": { "200": { "description": "Default Response" } }
      },
      "get": {
        "summary": "List lists",
        "tags": ["Lists"],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/lists/{id}": {
      "get": {
        "summary": "Get a list",
        "tags": ["Lists"],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "patch": {
        "summary": "Update a list",
        "tags": ["Lists"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "minLength": 1 },
                  "description": { "type": "string", "nullable": true },
                  "color": { "type": "string", "nullable": true }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "delete": {
        "summary": "Delete a list",
        "tags": ["Lists"],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/lists/{id}/members": {
      "post": {
        "summary": "Add a subscriber to a list",
        "tags": ["Lists"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "subscriberId": { "type": "string", "format": "uuid" } },
                "required": ["subscriberId"],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "get": {
        "summary": "List list members",
        "tags": ["Lists"],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0,
              "maximum": 1000
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": { "type": "integer", "minimum": 0 },
            "in": "query",
            "name": "offset",
            "required": false
          },
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/lists/{id}/members/{subscriberId}": {
      "delete": {
        "summary": "Remove a subscriber from a list",
        "tags": ["Lists"],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "subscriberId",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/segments": {
      "post": {
        "summary": "Create a segment",
        "tags": ["Segments"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "minLength": 1 },
                  "description": { "type": "string", "nullable": true },
                  "matchType": { "type": "string", "enum": ["all", "any"], "default": "all" },
                  "rules": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "field": { "type": "string", "minLength": 1 },
                        "op": {
                          "type": "string",
                          "enum": ["eq", "neq", "contains", "gt", "lt", "exists", "not_exists"]
                        },
                        "value": {}
                      },
                      "required": ["field", "op"],
                      "additionalProperties": false
                    },
                    "default": []
                  }
                },
                "required": ["name"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": { "200": { "description": "Default Response" } }
      },
      "get": {
        "summary": "List segments",
        "tags": ["Segments"],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/segments/{id}": {
      "get": {
        "summary": "Get a segment",
        "tags": ["Segments"],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "patch": {
        "summary": "Update a segment",
        "tags": ["Segments"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "minLength": 1 },
                  "description": { "type": "string", "nullable": true },
                  "matchType": { "type": "string", "enum": ["all", "any"], "default": "all" },
                  "rules": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "field": { "type": "string", "minLength": 1 },
                        "op": {
                          "type": "string",
                          "enum": ["eq", "neq", "contains", "gt", "lt", "exists", "not_exists"]
                        },
                        "value": {}
                      },
                      "required": ["field", "op"],
                      "additionalProperties": false
                    },
                    "default": []
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "delete": {
        "summary": "Delete a segment",
        "tags": ["Segments"],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/segments/preview": {
      "post": {
        "summary": "Preview an ad-hoc rule set: matching subscribers + total count",
        "tags": ["Segments"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "matchType": { "type": "string", "enum": ["all", "any"], "default": "all" },
                  "rules": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "field": { "type": "string", "minLength": 1 },
                        "op": {
                          "type": "string",
                          "enum": ["eq", "neq", "contains", "gt", "lt", "exists", "not_exists"]
                        },
                        "value": {}
                      },
                      "required": ["field", "op"],
                      "additionalProperties": false
                    },
                    "default": []
                  },
                  "limit": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0,
                    "maximum": 1000
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/segments/{id}/subscribers": {
      "get": {
        "summary": "Evaluate a saved segment",
        "tags": ["Segments"],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0,
              "maximum": 1000
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": { "type": "integer", "minimum": 0 },
            "in": "query",
            "name": "offset",
            "required": false
          },
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
...
    "/v1/stats/activity": {
      "get": {
        "summary": "Daily send activity, zero-filled (no open/click data yet)",
        "tags": ["Stats"],
        "parameters": [
          {
            "schema": { "type": "integer", "exclusiveMinimum": true, "minimum": 0, "maximum": 365 },
            "in": "query",
            "name": "days",
            "required": false
          },
          {
            "schema": { "type": "string", "enum": ["email", "sms", "whatsapp", "voice"] },
            "in": "query",
            "name": "channel",
            "required": false
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
...
  }
}